+++ src/power_pptx/__init__.py
@@ -55,7 +55,7 @@ from power_pptx.parts.slide import (
 if TYPE_CHECKING:
     from power_pptx.opc.package import Part
 
-__version__ = "2.8.1"
+__version__ = "2.9.0"
 
 sys.modules["power_pptx.exceptions"] = exceptions
 del sys